Creationimg/requirements.txt
streamlit
opencv-python-headless
numpy
Pillow
stripe
pymongo
bcrypt
python-dotenv